+++ dwds/lib/src/debugging/instance.dart
@@ -306,7 +306,8 @@ class InstanceHelper extends Domain {
     // We do this in in awkward way because we want the keys and values, but we
     // can't return things by value or some Dart objects will come back as
     // values that we need to be RemoteObject, e.g. a List of int.
-    final expression = _jsRuntimeFunctionCall('getMapElements(this)');
+    final expression = globalToolConfiguration.loadStrategy.dartRuntimeDebugger
+        .getMapElementsJsExpression();
 
     final keysAndValues = await inspector.jsCallFunctionOn(map, expression, []);
     final keys = await inspector.loadField(keysAndValues, 'keys');
@@ -523,7 +524,8 @@ class InstanceHelper extends Domain {
     // We do this in in awkward way because we want the keys and values, but we
     // can't return things by value or some Dart objects will come back as
     // values that we need to be RemoteObject, e.g. a List of int.
-    final expression = _jsRuntimeFunctionCall('getRecordFields(this)');
+    final expression = globalToolConfiguration.loadStrategy.dartRuntimeDebugger
+        .getRecordFieldsJsExpression();
 
     final result = await inspector.jsCallFunctionOn(record, expression, []);
     final fieldNameElements =
@@ -674,8 +676,8 @@ class InstanceHelper extends Domain {
     final length = metaData.length;
     final objectId = remoteObject.objectId;
     if (objectId == null) return null;
-
-    final expression = _jsRuntimeFunctionCall('getSetElements(this)');
+    final expression = globalToolConfiguration.loadStrategy.dartRuntimeDebugger
+        .getSetElementsJsExpression();
 
     final result =
         await inspector.jsCallFunctionOn(remoteObject, expression, []);