+++ dwds/test/instances/record_inspection_amd_canary_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
@@ -7,11 +7,12 @@
 @Timeout(Duration(minutes: 2))
 library;
 
+import 'package:dwds/src/services/expression_compiler.dart';
 import 'package:test/test.dart';
 import 'package:test_common/test_sdk_configuration.dart';
 
 import '../fixtures/context.dart';
-import 'common/instance_inspection_common.dart';
+import 'common/record_inspection_common.dart';
 
 void main() {
   // Enable verbose logging for debugging.
@@ -22,6 +23,7 @@ void main() {
     final provider = TestSdkConfigurationProvider(
       verbose: debug,
       canaryFeatures: canaryFeatures,
+      ddcModuleFormat: ModuleFormat.amd,
     );
     tearDownAll(provider.dispose);